% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupQueue.R
\name{BackupQueue}
\alias{BackupQueue}
\alias{BackupQueueIndex}
\alias{BackupQueueDateTime}
\alias{BackupQueueDate}
\title{An R6 class for managing backups}
\description{
\code{BackupQueue} & co are part of the \link[R6:R6]{R6} API of \strong{rotor}. They are
used internally by \code{\link[=rotate]{rotate()}} and related functions and are not designed
for interactive use. Rather, if you are a package developer and want to
integrate rotor in one of your package, the \code{BackupQueue} subclasses give
you a bit of extra control.

As of now, \strong{the R6 API is still under development and subject to change}.
More documentation will follow in the foreseable future.
}
\section{Usage}{
\preformatted{
x <- BackupQueueIndex$new(file, backup_dir = dirname(file), max_backups = Inf,
  compression = FALSE)

x <- BackupQueueDate$new(file, backup_dir = dirname(file), max_backups = Inf,
  compression = FALSE, fmt = "\%Y-\%m-\%d", cache_backups = FALSE)

x <- BackupQueueDateTime$new(file, backup_dir = dirname(file), max_backups =
  Inf, compression = FALSE, fmt = "\%Y-\%m-\%d--\%H-\%M-\%S", cache_backups = FALSE)

x$increment_index(n = 1)
x$pad_index()
x$print()
x$prune(max_backups = self$max_backups)
x$push_backup()
x$push_backup(overwrite = FALSE, now = Sys.time())
x$set_backup_dir(x)
x$set_cache_backups(x)
x$set_compression(x)
x$set_file(x)
x$set_fmt(x)
x$set_max_backups(x)
x$should_rotate(size, age, now = Sys.time(), last_rotation = self$last_rotation)
x$should_rotate(size)
x$update_backups_cache()

x$backup_dir
x$backups
x$cache_backups
x$compression
x$file
x$fmt
x$has_backups
x$last_rotation
x$max_backups
x$n_backups

}
}

