% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupQueue.R
\name{BackupQueueIndex}
\alias{BackupQueueIndex}
\title{An R6 class for managing indexed backups}
\description{
An R6 class for managing indexed backups

An R6 class for managing indexed backups
}
\details{
As of now, \strong{the R6 API is still experimental and subject to change}.
}
\seealso{
Other BackupQueue: 
\code{\link{BackupQueueDateTime}},
\code{\link{BackupQueue}}
}
\concept{BackupQueue}
\section{Super class}{
\code{\link[rotor:BackupQueue]{rotor::BackupQueue}} -> \code{BackupQueueIndex}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-push}{\code{BackupQueueIndex$push()}}
\item \href{#method-prune}{\code{BackupQueueIndex$prune()}}
\item \href{#method-should_rotate}{\code{BackupQueueIndex$should_rotate()}}
\item \href{#method-pad_index}{\code{BackupQueueIndex$pad_index()}}
\item \href{#method-increment_index}{\code{BackupQueueIndex$increment_index()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="BackupQueue" data-id="initialize">}\href{../../rotor/html/BackupQueue.html#method-initialize}{\code{rotor::BackupQueue$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="BackupQueue" data-id="print">}\href{../../rotor/html/BackupQueue.html#method-print}{\code{rotor::BackupQueue$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="BackupQueue" data-id="set_compression">}\href{../../rotor/html/BackupQueue.html#method-set_compression}{\code{rotor::BackupQueue$set_compression()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="BackupQueue" data-id="set_dir">}\href{../../rotor/html/BackupQueue.html#method-set_dir}{\code{rotor::BackupQueue$set_dir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="BackupQueue" data-id="set_file">}\href{../../rotor/html/BackupQueue.html#method-set_file}{\code{rotor::BackupQueue$set_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="BackupQueue" data-id="set_max_backups">}\href{../../rotor/html/BackupQueue.html#method-set_max_backups}{\code{rotor::BackupQueue$set_max_backups()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
Create a new index-stamped backup (e.g. \file{logfile.1.log})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueueIndex$push()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prune"></a>}}
\if{latex}{\out{\hypertarget{method-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueueIndex$prune(max_backups = self$max_backups)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-should_rotate"></a>}}
\if{latex}{\out{\hypertarget{method-should_rotate}{}}}
\subsection{Method \code{should_rotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueueIndex$should_rotate(size, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} or \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pad_index"></a>}}
\if{latex}{\out{\hypertarget{method-pad_index}{}}}
\subsection{Method \code{pad_index()}}{
Pad the indices in the filenames of indexed backups
to the number of digits of the largest index. Usually does not have to
be called manually.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueueIndex$pad_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-increment_index"></a>}}
\if{latex}{\out{\hypertarget{method-increment_index}{}}}
\subsection{Method \code{increment_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueueIndex$increment_index(n = 1)}\if{html}{\out{</div>}}
}

}
}
