% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{DirectoryQueue}
\alias{DirectoryQueue}
\title{An R6 class for managing a persistent file-based cache}
\description{
\code{Cache} provides an \link[R6:R6Class]{R6} API for managing a folder that
contains cached \R objects as \code{.rds} files. \code{Cache} supports
automatic removal of old files if the the cache exceeds a predetermined
number of files, total size, or if the files exceed a certain age.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dir}}{a \code{character} scalar. path of the directory in which to store
the cache files}

\item{\code{max_files}}{see the \code{compress} argument of \code{\link[base:readRDS]{base::saveRDS()}}.
\strong{Note}: this differs from the \verb{$compress} argument of \code{\link[=rotate]{rotate()}}.}

\item{\code{max_files}}{\code{integer} scalar: maximum number of files to keep in
the cache}

\item{\code{max_size}}{scalar \code{integer}, \code{character} or \code{Inf}. Delete
cached files (starting with the oldest) until the total size of the
cache is below \code{max_size}. \code{Integers} are interpreted as bytes. You
can pass \code{character} vectors that contain a file size suffix like \verb{1k}
(kilobytes), \verb{3M} (megabytes), \verb{4G} (gigabytes), \verb{5T} (terabytes). Instead
of these short forms you can also be explicit and use the IEC suffixes
\code{KiB}, \code{MiB}, \code{GiB}, \code{TiB}. In Both cases \code{1} kilobyte is \code{1024} bytes, 1
\code{megabyte} is \code{1024} kilobytes, etc... .}

\item{\code{max_age}}{\itemize{
\item a \code{Date} scalar: Remove all backups before this date
\item a \code{character} scalar representing a Date in ISO format (e.g. \code{"2019-12-31"})
\item a \code{character} scalar representing an Interval in the form \code{"<number> <interval>"} (see \code{\link[=rotate]{rotate()}})
}}

\item{\code{hashfun}}{\code{NULL} or a \code{function} to generate a unique hash from the
object to be cached (see example). The hash \emph{must} be a text string
that is a valid filename on the target system. If \verb{$hashfun} is \code{NULL},
a storage key must be supplied manually in \code{cache$push()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DirectoryQueue$new()}}
\item \href{#method-push}{\code{DirectoryQueue$push()}}
\item \href{#method-remove}{\code{DirectoryQueue$remove()}}
\item \href{#method-pop}{\code{DirectoryQueue$pop()}}
\item \href{#method-prune}{\code{DirectoryQueue$prune()}}
\item \href{#method-purge}{\code{DirectoryQueue$purge()}}
\item \href{#method-set_dir}{\code{DirectoryQueue$set_dir()}}
\item \href{#method-set_max_files}{\code{DirectoryQueue$set_max_files()}}
\item \href{#method-set_max_age}{\code{DirectoryQueue$set_max_age()}}
\item \href{#method-set_max_size}{\code{DirectoryQueue$set_max_size()}}
\item \href{#method-set_compression}{\code{DirectoryQueue$set_compression()}}
\item \href{#method-set_hashfun}{\code{DirectoryQueue$set_hashfun()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$push(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$remove(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\if{latex}{\out{\hypertarget{method-pop}{}}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$pop(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prune"></a>}}
\if{latex}{\out{\hypertarget{method-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$prune(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-purge"></a>}}
\if{latex}{\out{\hypertarget{method-purge}{}}}
\subsection{Method \code{purge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$purge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_dir"></a>}}
\if{latex}{\out{\hypertarget{method-set_dir}{}}}
\subsection{Method \code{set_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_dir(x, create = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_files"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_files}{}}}
\subsection{Method \code{set_max_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_max_files(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_age"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_age}{}}}
\subsection{Method \code{set_max_age()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_max_age(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_size"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_size}{}}}
\subsection{Method \code{set_max_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_max_size(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_compression"></a>}}
\if{latex}{\out{\hypertarget{method-set_compression}{}}}
\subsection{Method \code{set_compression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_compression(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_hashfun"></a>}}
\if{latex}{\out{\hypertarget{method-set_hashfun}{}}}
\subsection{Method \code{set_hashfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_hashfun(x)}\if{html}{\out{</div>}}
}

}
}
