% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotor-package.R
\docType{package}
\name{rotor-package}
\alias{rotor}
\alias{rotor-package}
\title{rotor: Rote files}
\description{
Rotate or backup files based on size or date. Inspired the linux 
  utility logrotate.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (0000-0003-3344-9851)

}
\keyword{internal}
