% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupQueue.R
\name{BackupQueue}
\alias{BackupQueue}
\title{An R6 class for managing backups}
\description{
\code{BackupQueue} & co are part of the \link[R6:R6Class]{R6} API of \strong{rotor}. They
are used internally by \code{\link[=rotate]{rotate()}} and related functions and are not designed
for interactive use. Rather, if you are a package developer and want to
integrate rotor in one of your package, the \code{BackupQueue} subclasses give you
a bit of extra control.
}
\seealso{
Other BackupQueue: 
\code{\link{BackupQueueDateTime}},
\code{\link{BackupQueueIndex}}
}
\concept{BackupQueue}
\section{Super class}{
\code{\link[rotor:DirectoryQueue]{rotor::DirectoryQueue}} -> \code{BackupQueue}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dir}}{\code{character} scalar. Directory in which to place the backups.}

\item{\code{n}}{\code{integer} scalar. The number of backups that exist for \code{BackupQueue$file} sdfgs}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dir}}{\code{character} scalar. Directory in which to place the backups.}

\item{\code{n}}{\code{integer} scalar. The number of backups that exist for \code{BackupQueue$file} sdfgs}

\item{\code{file}}{\code{character} scalar. The file to backup/rotate.}

\item{\code{compression}}{(Optional) compression to use \code{compression} argument of \code{\link[=rotate]{rotate()}}.}

\item{\code{max_backups}}{Maximum number/size/age of backups. See \code{max_backups}
argument of \code{\link[=rotate]{rotate()}}}

\item{\code{has_backups}}{Returns \code{TRUE} if at least one backup of \code{BackupQueue$file}
exists
All backups of self$file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BackupQueue$new()}}
\item \href{#method-prune}{\code{BackupQueue$prune()}}
\item \href{#method-print}{\code{BackupQueue$print()}}
\item \href{#method-set_file}{\code{BackupQueue$set_file()}}
\item \href{#method-set_compression}{\code{BackupQueue$set_compression()}}
\item \href{#method-set_max_backups}{\code{BackupQueue$set_max_backups()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="DirectoryQueue" data-id="push">}\href{../../rotor/html/DirectoryQueue.html#method-push}{\code{rotor::DirectoryQueue$push()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rotor" data-topic="DirectoryQueue" data-id="set_dir">}\href{../../rotor/html/DirectoryQueue.html#method-set_dir}{\code{rotor::DirectoryQueue$set_dir()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueue$new(
  file,
  dir = dirname(file),
  max_backups = Inf,
  compression = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prune"></a>}}
\if{latex}{\out{\hypertarget{method-prune}{}}}
\subsection{Method \code{prune()}}{
Delete all backups except \code{max_backups}. See \code{\link[=prune_backups]{prune_backups()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueue$prune(max_backups = self$max_backups)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueue$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_file"></a>}}
\if{latex}{\out{\hypertarget{method-set_file}{}}}
\subsection{Method \code{set_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueue$set_file(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_compression"></a>}}
\if{latex}{\out{\hypertarget{method-set_compression}{}}}
\subsection{Method \code{set_compression()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueue$set_compression(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_backups"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_backups}{}}}
\subsection{Method \code{set_max_backups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackupQueue$set_max_backups(x)}\if{html}{\out{</div>}}
}

}
}
