% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupQueue.R
\docType{data}
\name{BackupQueue}
\alias{BackupQueue}
\alias{BackupQueueIndex}
\alias{BackupQueueDateTime}
\alias{BackupQueueDate}
\title{An R6 class for managing backups}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BackupQueue

BackupQueueIndex

BackupQueueDateTime

BackupQueueDate
}
\description{
\code{BackupQueue} & co are part of the \link[R6:R6]{R6} API of \strong{rotor}. They are
used internally by \code{\link[=rotate]{rotate()}} and related functions and are not designed
for interactive use. Rather, if you are a package developer and want to
integrate rotor in one of your package, the \code{BackupQueue} subclasses give
you a bit of extra control.

As of now, \strong{the R6 API is still under development and subject to change}.
More documentation will follow in the foreseable future.
}
\keyword{datasets}
