% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupQueue.R, R/BackupQueueDate.R,
%   R/BackupQueueDateTime.R, R/BackupQueueIndex.R
\docType{data}
\name{BackupQueue}
\alias{BackupQueue}
\alias{BackupQueueDate}
\alias{BackupQueueDateTime}
\alias{BackupQueueIndex}
\title{An internal R6 class for managing backups}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BackupQueue

BackupQueueDate

BackupQueueDateTime

BackupQueueIndex
}
\description{
\strong{documentation under development}
\code{BackupQueue} and its subclasses are \link[R6:R6Class]{R6::R6Class} for rotating files and
managing backup. They are the internal constructs on which \code{\link[=rotate]{rotate()}} and
co. are based. The \code{BackupQueue} constructor is exported for use by other
package developers and not intended for direct use.
}
\keyword{datasets}
