% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectoryQueue.R
\name{DirectoryQueue}
\alias{DirectoryQueue}
\title{An R6 class for managing a persistent file-based queue}
\description{
An R6 class for managing a persistent file-based queue

An R6 class for managing a persistent file-based queue
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dir}}{a \code{character} scalar. path of the directory in which to store
the cache files}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DirectoryQueue$new()}}
\item \href{#method-push}{\code{DirectoryQueue$push()}}
\item \href{#method-prune}{\code{DirectoryQueue$prune()}}
\item \href{#method-set_dir}{\code{DirectoryQueue$set_dir()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$push(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prune"></a>}}
\if{latex}{\out{\hypertarget{method-prune}{}}}
\subsection{Method \code{prune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$prune(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_dir"></a>}}
\if{latex}{\out{\hypertarget{method-set_dir}{}}}
\subsection{Method \code{set_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectoryQueue$set_dir(x, create = TRUE)}\if{html}{\out{</div>}}
}

}
}
