% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotor-package.R
\docType{package}
\name{rotor-package}
\alias{rotor}
\alias{rotor-package}
\title{rotor: Log Rotation and Conditional Backups}
\description{
Conditionally rotate or back-up files based on their size or the
  date of the last backup; inspired by the 'linux' utility 'logrotate'.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (0000-0003-3344-9851)

}
\keyword{internal}
